What you can configure Users collection path The path to the Firestore collection where user documents will be created.
Fields to populate
Select the fields from the user record that you want to populate in the Firestore document. If you select a field that does not exist in the user record, it will be ignored.

Delete document on user delete
If you enable this option, the extension will automatically delete the Firestore user document when the user is deleted from Firebase Authentication.

Backfill existing users
If you enable this option, the extension will create Firestore user documents for all existing users right after installation is complete.

Cloud Functions location
Where do you want to deploy the functions created for this extension? For help selecting a location, refer to the location selection guide. Cloud functions
createUserDocument
A function that is triggered when a new user is created in Firebase Authentication. It creates a user document in a specified Firestore collection and populates it with the selected fields from the user record.
deleteUserDocument
A function that is triggered when a user is deleted from Firebase Authentication. If the instance is configured to do so, it deletes the user document from the specified Firestore collection.
backfillExistingUsers
A function that is triggered right after the extension installation is complete. If the instance is configured to do so, it creates user documents in the specified Firestore collection for all existing users.